+++ src/hashable.rs
@@ -1,0 +1,5 @@
+use crate::Hash;
+
+pub trait Hashable {
+    fn hash(&self) -> Hash;
+}