+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## OxChain 0.1.3 (2022-01-04)
+
+### Added
+
+- Added `Hashable` trait.
+    - Implemented for `Block` and `Transaction`.
+- Added a new `application` module.
+- Added a new test case for `find_merkle_root` function.
+
+### Removed
+
+- Removed `calculate_hash` function in `Block`.
+
 ## OxChain 0.1.2 (2022-12-29)
 
 ### Changed